Final updates and fixing bugs

+++ app/routers/user.py
@@ -58,8 +58,8 @@ def get_all_users(db: Session = Depends(get_db),
 
 # Get current user personal info
 
-@router.get('/info/{id}',response_model=schemas.UserDetails)
-def get_current_user_information(id:int, user_pwd: schemas.UserPWD, db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
+@router.get('/info/{id}',response_model=schemas.UserOut)
+def get_current_user_information(id:int, db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
 
     user= db.query(models.User).filter(models.User.id==id).first()
     #add code for dehashing the password
@@ -70,14 +70,10 @@ def get_current_user_information(id:int, user_pwd: schemas.UserPWD, db: Session 
     if user.id!=current_user.id:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
 
-    if verify(user_pwd.password, user.password):
-        # Create a new instance of UserDetails with the dehashed password
-        user_details = schemas.UserDetails(**user.__dict__)
-        user_details.password = user_pwd.password
+    user_details = schemas.UserOut(**user.__dict__)
 
-        return user_details
+    return user_details
 
-    raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid password")
 
 
 
@@ -85,13 +81,16 @@ def get_current_user_information(id:int, user_pwd: schemas.UserPWD, db: Session 
 # Delete user endpoint
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(id: int,db: Session = Depends(get_db),current_admin: int = Depends(oauth2.get_current_admin)):
+def delete_user(id: int,db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
 
     user_query=db.query(models.User).filter(models.User.id==id)
     user_del=user_query.first()
 
     if user_del==None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"user with id: {id} does not exist")
+    
+    if user_del.id!=current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
 
     user_query.delete(synchronize_session=False)
     db.commit()
